requirements.txt
attrs==25.3.0
certifi==2025.4.26
cff==1.0.0a1
cffconvert==2.0.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
convert==0.1.2
cycler==0.12.1
docopt==0.6.2
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
grad-cam==1.5.5
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==3.2.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pykwalify==1.8.0
pyparsing==3.2.3
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
pytorch-gradcam==0.2.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.0.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
ttach==0.0.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
